+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Patched
+- Also set timestamp and buffer duration on per-frame metadata buffers.
+
 ## [0.1.5] - 2019-10-30
 
 ### Patched